requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements.txt pyproject.toml
#
certifi==2025.8.3
    # via sentry-sdk
dnspython==2.7.0
    # via pymongo
grpcio==1.74.0
    # via
    #   grpcio-tools
    #   shield-receiver (pyproject.toml)
grpcio-tools==1.74.0
    # via shield-receiver (pyproject.toml)
protobuf==6.31.1
    # via
    #   grpcio-tools
    #   shield-receiver (pyproject.toml)
psycopg2-binary==2.9.10
    # via shield-receiver (pyproject.toml)
pymongo==4.13.2
    # via shield-receiver (pyproject.toml)
python-dotenv==1.1.1
    # via shield-receiver (pyproject.toml)
sentry-sdk==2.34.1
    # via shield-receiver (pyproject.toml)
urllib3==2.5.0
    # via sentry-sdk

# The following packages are considered to be unsafe in a requirements file:
# setuptools
